services/detector/pyproject.toml
[project]
name = "detector"
version = "0.1.0"
description = ""
authors = [
    {name = "ycalk",email = "alosev752@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "ai-image-detector-messaging-schema (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)"
]

[tool.poetry]
packages = [{include = "detector", from = "src"}]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies]
torch = {source = "pytorch"}
torchvision = {source = "pytorch"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
detector = "detector:run"